Depth Texture added to Graphics Pipeline

+++ src/KTXCompressor/KTXCompressor/Renderer/Presentation/FrameBuffer.cpp
@@ -15,7 +15,7 @@ namespace KTXCompressor {
         VkFramebufferCreateInfo framebufferCreateInfo = {};
         framebufferCreateInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
         framebufferCreateInfo.renderPass = vulkanRenderPass;
-        framebufferCreateInfo.attachmentCount = attachments.size();
+        framebufferCreateInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
         framebufferCreateInfo.pAttachments = attachments.data();
         framebufferCreateInfo.width = extent.width;
         framebufferCreateInfo.height = extent.height;
